requirements.txt
streamlit==1.33.0
pandas==2.2.1
python-dotenv==1.0.1
pypdf==4.1.0
langchain==0.1.16
openai==1.16.2
chromadb==0.4.24
langchain-openai==0.1.3
langchain-community==0.0.34
pysqlite3-binary==0.5.2.1
onnxruntime==1.17.1